// MobileAPP/src/utilis/types/Registration.ts
export interface registerSecondValueType {
  raasiAndMoonSign: any;
  Star: any;
  dosham: any;
  doshamYes: any;
  country: any;
  city: any;
  state: any;
  Hour: any;
  Minutes: any;
  AMAndPM: any;
  doshamError: string;
  doshamYesError: string;
  raasiAndMoonSignError: string;
  StarError: string;
  countryError: any;
  cityError: string;
  stateError: string;
  HourError: string;
  MinutesError: string;
  AMAndPMError: any;
}

export const registerSecondFormValue = {
  raasiAndMoonSign: '',
  Star: '',
  dosham: '',
  doshamYes: '',
  country: '',
  city: '',
  state: '',
  Hour: '',
  Minutes: '',
  AMAndPM: '',
  raasiAndMoonSignError: '',
  StarError: '',
  doshamError: '',
  doshamYesError: '',
  countryError: '',
  cityError: '',
  stateError: '',
  HourError: '',
  MinutesError: '',
  AMAndPMError: '',
};